pyproject.toml
[build-system]
requires = ["setuptools>=69.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "ufabc-chatbot"
version = "0.1.0"
description = "Initial architecture scaffold for the UFABC chatbot project."
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "fastapi>=0.115.0",
  "uvicorn[standard]>=0.30.0",
  "pydantic-settings>=2.4.0",
  "sqlalchemy>=2.0.36",
  "greenlet>=3.0.0",
  "asyncpg>=0.29.0",
  "aiosqlite>=0.20.0",
  "boto3>=1.35.0",
  "pyyaml>=6.0.0",
  "python-multipart>=0.0.9",
  "openai>=1.0.0",
  "bcrypt>=4.0.0",
  "python-jose[cryptography]>=3.3.0",
  "pydantic[email]>=2.0.0",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.3.0",
  "httpx>=0.27.0",
]

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
pythonpath = ["src"]
